Add property intelligence features, checklist system, and comprehensive test suite
Batch 1 implementation: property location fields, 18-item checklist with auto-assessment,
property detail page with map embed and data panels, dashboard saved property cards,
side-by-side comparison page, address lookup service, and EPC API improvements.
Adds Telescope, Pulse, Sanctum, Scout, Dusk infrastructure. Removes Horizon/Redis dependency.
406 tests, 838 assertions all passing.

+++ app/Services/Api/LandRegistryApiService.php
@@ -72,7 +72,7 @@ class LandRegistryApiService implements PropertyDataProvider
             return true;
         }
 
-        $ttl = config('housescout.api.land_registry.cache_ttl');
+        $ttl = (int) config('housescout.api.land_registry.cache_ttl');
 
         return $landData->fetched_at->addSeconds($ttl)->isPast();
     }
